Moved Readme unit tests to separate file and added additional tests
