+++ tests/unit/test_model.py
@@ -1356,61 +1356,3 @@ def test_create_readme_attributes_without_required_parameters_raises_exception(
 ):
     with pytest.raises(error, match=message):
         Readme.Attributes.create(asset=asset, content=content, asset_name=asset_name)
-
-
-@pytest.mark.parametrize(
-    "asset, content, asset_name, error, message",
-    [
-        (None, "stuff", None, ValueError, "asset is required"),
-        (
-            Table.create(
-                name=TABLE_NAME,
-                schema_qualified_name=SCHEMA_QUALIFIED_NAME,
-            ),
-            None,
-            None,
-            ValueError,
-            "content is required",
-        ),
-        (
-            Table(),
-            "stuff",
-            None,
-            ValueError,
-            "asset_name is required when name is not available from asset",
-        ),
-    ],
-)
-def test_create_readme_without_required_parameters_raises_exception(
-    asset, content, asset_name, error, message
-):
-    with pytest.raises(error, match=message):
-        Readme.create(asset=asset, content=content, asset_name=asset_name)
-
-
-@pytest.mark.parametrize(
-    "asset, content, asset_name, expected_name",
-    [
-        (
-            Table.create(
-                name=TABLE_NAME,
-                schema_qualified_name=SCHEMA_QUALIFIED_NAME,
-            ),
-            "<h1>stuff</h1>",
-            None,
-            TABLE_NAME,
-        ),
-        (
-            Table(attributes=Table.Attributes()),
-            "<h1>stuff</h1>",
-            TABLE_NAME,
-            TABLE_NAME,
-        ),
-    ],
-)
-def test_create_readme(asset, content, asset_name, expected_name):
-    readme = Readme.create(asset=asset, content=content, asset_name=asset_name)
-    assert readme.qualified_name == f"{asset.guid}/readme"
-    assert readme.name == f"{expected_name} Readme"
-    assert readme.attributes.asset == asset
-    assert readme.description == content
